disabled DashboardTest and DashboardWidgetTest, changed list of permissions, moved workItems tests to new agent APIs
