+++ src/test/java/com/qaprosoft/zafira/api/TestV1Test.java
@@ -2,7 +2,7 @@ package com.qaprosoft.zafira.api;
 
 import com.qaprosoft.apitools.validation.JsonCompareKeywords;
 import com.qaprosoft.carina.core.foundation.utils.R;
-import com.qaprosoft.zafira.api.testController.GetTestResultsHistoryByIdMethod;
+import com.qaprosoft.zafira.api.testController.*;
 import com.qaprosoft.zafira.api.testRunController.v1.DeleteTestByIdV1Method;
 import com.qaprosoft.zafira.api.testRunController.v1.GetTestsByCiRunIdV1Method;
 import com.qaprosoft.zafira.api.testRunController.v1.PostStartTestsInTestRunV1Method;
@@ -26,6 +26,7 @@ import org.testng.annotations.DataProvider;
 import org.testng.annotations.Test;
 
 import java.time.OffsetDateTime;
+import java.util.List;
 
 public class TestV1Test extends ZafiraAPIBaseTest {
 
@@ -328,6 +329,80 @@ public class TestV1Test extends ZafiraAPIBaseTest {
         apiExecutor.expectStatus(getTestResultsHistoryByIdMethod, HTTPStatusCodeType.OK);
         apiExecutor.validateResponse(getTestResultsHistoryByIdMethod, JSONCompareMode.STRICT,
                 JsonCompareKeywords.ARRAY_CONTAINS.getKey());
+    }
+
+    @Test
+    public void testLinkWorkItemToTest() {
+        String expectedJiraIdValue = R.TESTDATA.get(ConfigConstant.EXPECTED_JIRA_ID_KEY);
+        String workItemType = R.TESTDATA.get(ConfigConstant.WORK_ITEM_TYPE_KEY);
+        testRunId = startTestRunV1();
+        int testId=startTestV1(testRunId);
+        new TestServiceV1Impl().finishTestAsResult(testRunId,testId,"FAILED");
+        PostLinkWorkItemMethod postLinkWorkItemMethod = new PostLinkWorkItemMethod(1, expectedJiraIdValue,
+                testId, workItemType);
+        apiExecutor.expectStatus(postLinkWorkItemMethod, HTTPStatusCodeType.OK);
+        String linkWorkItemRs = apiExecutor.callApiMethod(postLinkWorkItemMethod);
+        String jiraId = JsonPath.from(linkWorkItemRs).get(JSONConstant.JIRA_ID_KEY);
+        Assert.assertEquals(jiraId, expectedJiraIdValue, "Work item was not linked to test!");
+    }
+
+    @Test
+    public void testLinkWorkItemToTestWithStatusPASSED() {
+        String expectedJiraIdValue = R.TESTDATA.get(ConfigConstant.EXPECTED_JIRA_ID_KEY);
+        String workItemType = R.TESTDATA.get(ConfigConstant.WORK_ITEM_TYPE_KEY);
+        testRunId = startTestRunV1();
+        int testId=startTestV1(testRunId);
+        new TestServiceV1Impl().finishTestAsResult(testRunId,testId,"PASSED");
+        PostLinkWorkItemMethod postLinkWorkItemMethod = new PostLinkWorkItemMethod(1, expectedJiraIdValue,
+                testId, workItemType);
+        apiExecutor.expectStatus(postLinkWorkItemMethod, HTTPStatusCodeType.FORBIDDEN);
+    }
+
+    @Test
+    public void testGetWorkItem() {
+        String workItemType = R.TESTDATA.get(ConfigConstant.WORK_ITEM_TYPE_KEY);
+        String jiraId = R.TESTDATA.get(ConfigConstant.EXPECTED_JIRA_ID_KEY);
+        testRunId = startTestRunV1();
+        int testId=startTestV1(testRunId);
+        new TestServiceV1Impl().finishTestAsResult(testRunId,testId,"FAILED");
+        apiExecutor.callApiMethod(new PostLinkWorkItemMethod(1, jiraId, testId, workItemType));
+        GetWorkItemMethod getWorkItemMethod = new GetWorkItemMethod(testId, workItemType);
+        apiExecutor.expectStatus(getWorkItemMethod, HTTPStatusCodeType.OK);
+        apiExecutor.callApiMethod(getWorkItemMethod);
+        apiExecutor.validateResponse(getWorkItemMethod,
+                JSONCompareMode.LENIENT, JsonCompareKeywords.ARRAY_CONTAINS.getKey());
+    }
+
+    @Test
+    public void testCreateBatchWorkItem() {
+        String jiraId = R.TESTDATA.get(ConfigConstant.EXPECTED_JIRA_ID_KEY);
+        testRunId = startTestRunV1();
+        int testId=startTestV1(testRunId);
+        new TestServiceV1Impl().finishTestAsResult(testRunId,testId,"FAILED");
+        PostCreateBatchWorkItemsMethod postCreateBatchWorkItemsMethod = new PostCreateBatchWorkItemsMethod(testRunId,testId, jiraId);
+        apiExecutor.expectStatus(postCreateBatchWorkItemsMethod, HTTPStatusCodeType.OK);
+        apiExecutor.callApiMethod(postCreateBatchWorkItemsMethod);
+        apiExecutor.validateResponse(postCreateBatchWorkItemsMethod, JSONCompareMode.STRICT, JsonCompareKeywords.ARRAY_CONTAINS.getKey());
+    }
+
+    @Test
+    public void testDeleteWorkItem() {
+        TestServiceImpl testServiсeImpl =  new TestServiceImpl();
+        String expectedJiraIdValue = R.TESTDATA.get(ConfigConstant.EXPECTED_JIRA_ID_KEY);
+        String workItemType = R.TESTDATA.get(ConfigConstant.WORK_ITEM_TYPE_KEY);
+        testRunId = startTestRunV1();
+        int testId=startTestV1(testRunId);
+        new TestServiceV1Impl().finishTestAsResult(testRunId,testId,"FAILED");
+        String linkWorkItemRs = apiExecutor.callApiMethod(new PostLinkWorkItemMethod(1, expectedJiraIdValue,
+                testId, workItemType));
+        int workItemId = JsonPath.from(linkWorkItemRs).get(JSONConstant.ID_KEY);
+        String testRs =testServiсeImpl.getAllTest(testRunId);
+        int workItemIdRs = JsonPath.from(testRs).get(JSONConstant.WORK_ITEM_ID_CHECK_KEY);
+        Assert.assertNotEquals(0, workItemIdRs, "Work item was not linked!");
+        apiExecutor.callApiMethod(new DeleteWorkItemMethod(testId, workItemId));
+        String testRsAfterDelete = testServiсeImpl.getAllTest(testRunId);
+        List<Integer> workItemsAfterDelete = JsonPath.from(testRsAfterDelete).getList(JSONConstant.WORK_ITEMS_ARRAY_KEY);
+        Assert.assertTrue(workItemsAfterDelete.isEmpty(), "Work item was not deleted!");
     }
 
 }
