+++ src/test/resources/api/permissionIAM/_get/rs.json
@@ -3,6 +3,8 @@
   "billing:payment-methods:update",
   "billing:read",
   "configuration:sso:saml",
+  "iam:api-tokens:read",
+  "iam:api-tokens:revoke",
   "iam:groups:delete",
   "iam:groups:read",
   "iam:groups:update",
@@ -16,23 +18,14 @@
   "projects:read",
   "projects:update",
   "reporting:dashboards:access-private",
-  "reporting:dashboards:delete",
-  "reporting:dashboards:update",
   "reporting:filters:create-public",
   "reporting:filters:delete-any",
+  "reporting:integrations:access-private",
   "reporting:integrations:read",
   "reporting:integrations:update",
-  "reporting:launchers:delete",
-  "reporting:launchers:read",
-  "reporting:launchers:update",
-  "reporting:settings:update",
   "reporting:stacktrace-labels:assign",
   "reporting:test-runs:access-private",
-  "reporting:test-runs:delete",
-  "reporting:test-runs:read",
-  "reporting:test-runs:update",
   "reporting:test-sessions:token:refresh",
-  "reporting:tests:update",
   "reporting:widgets:delete",
   "reporting:widgets:update"
 ]
