+++ src/test/resources/api/test/_put/rq.json
@@ -11,7 +11,7 @@
 "upstreamJobBuildNumber": 0,
 "buildNumber": 0,
 "startedBy": "HUMAN",
-"knownIssue": true,
+"knownIssue": false,
 "blocker": true,
 "reviewed": false
 }
