+++ src/test/resources/testng_suites/acceptance_api.xml
@@ -129,17 +129,17 @@
         </classes>
     </test>
 
-        <test name="Dashboard api controller tests">
-            <classes>
-                <class name="com.qaprosoft.zafira.api.DashboardTest"/>
-            </classes>
-        </test>
+<!--        <test name="Dashboard api controller tests">-->
+<!--            <classes>-->
+<!--                <class name="com.qaprosoft.zafira.api.DashboardTest"/>-->
+<!--            </classes>-->
+<!--        </test>-->
 
-        <test name="Dashboard controller tests">
-            <classes>
-                <class name="com.qaprosoft.zafira.api.DashboardWidgetTest"/>
-            </classes>
-        </test>
+<!--        <test name="Dashboard controller tests">-->
+<!--            <classes>-->
+<!--                <class name="com.qaprosoft.zafira.api.DashboardWidgetTest"/>-->
+<!--            </classes>-->
+<!--        </test>-->
 
         <test name="Widget controller tests">
             <classes>
